feat(syntax) change {template} to <template>
First of the steps needed for the HTML syntax introduction
as proposed in https://github.com/ariatemplates/hashspace/wiki/HTML-Syntax-Proposal

This just:

1) changes `{template}` syntax to `<template>`
2) removes the support for the old `# template` syntax

(all changes to the parser, samples, docs and test are all in this
commit).

The rest of the changes (introducing `<script>` for JS part, making
`<template>` accept HTMLish attribs) are to come.